// src/enums/GameEvents.ts
export enum GameEvent {
    goToLevel = 'goToLevel',
    timeLock = 'timeLock',
    kill = 'kill',
    startDialog = 'startDialog',
    endDialog = 'endDialog',
    restartLevel = 'restartLevel',
    onPlatform = 'onPlatform',
    offPlatform = 'offPlatform',
    grabObject = 'grabObject',
    throwObject = 'throwObject',
    releaseObject = 'releaseObject',
    destroyGame = 'destroyGame',
  }
  